add modularity for memory operation
